move request blocker, add test
